+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0
+- support `pantheon` agent type
+  - `code-tee pantheon dev ...`
+  - `code-tee pantheon review ...`
+- deprecate `panthon-tdd` agent type
+
 ## 1.1.0
 
 - support `--exec-path` to config agent bin path